Add logic for assigning unique PIDs to new processes/threads
We maintain an array "pidTable" which keeps track of all the assigned
PID values. For assigning a new PID, the least free PID is chosen which
is stored in "minFreePID". When a process/thread exits, "pidTable" and
other variables are appropriately updated.

+-----------+-----------------+
| Operation | Time complexity |
|           |    (average)    |
+-----------+-----------------+
| Insertion |       O(1)      |
+-----------+-----------------+
| Deletion  |       O(1)      |
+-----------+-----------------+

Currently the number of entries in "pidTable" is stored in the macro
THREADLIMIT which is 256 ; this is subjected to change.
It is currently debatable as to what should be the minimum value of the
PID that can be assigned to a new process/thread ; in "real" systems the
swapper has PID 0 and init has PID 1, however it is not yet confirm if
these exist currently in NachOS during runtime.

Test Plan:
```
./nachos -x ../test/testexec
```
The first line of output of above execution should be -
PID of the current process: 1