Fix incorrect parsing of executable name from memory in Syscall_Exec
TODO
----
The weird thing about this is that the exact same code for Syscall_Exec
(prior to the changes made in this commit) works totally fine. For an
instant I'd have thought that the problem was due to a low value of 64
bytes allocated to the buffer but that is not the case as the erroneous
behavior can be reproduced with 1024 bytes too. I remember coming across
the same anomaly when was trying to implement Syscall_Fork for the first
time, and that resulted in the exact same behavior. So it seems that
implementation of Syscall_Fork somehow effects how Syscall_Exec behaves.
Need to dig in more!