Fix calculation of CPU bursts
Earlier, the end of CPU burst was being marked inside ScheduleThread().
This will lead to erroneous increment in ticks of CPU bursts as the
thread can get stuck before calling ScheduleThread() in the while loop
waiting for an interrupt to wake it up (this is the case when the thread
moves to SLEEP state). Hence, we mark the end of CPU burst as soon as
the thread enters the SLEEP state or enter YieldCPU().

Fixes #1